refactor: moved translate to fucntions
This is the start of the refactor. The purpose is to store every
function in the special file and then call them. Thus, we do not need to
store 4 files for different functions - we store one, and then build on
top of that. This will also allow us to create a "main endpoint" of
speech-to-speech with the mode specification (either a conversation or a
translation)

Signed-off-by: paul <112290055+2560p@users.noreply.github.com>